func/requirements.txt
WebOb==1.8.5
requests==2.21.0
locustio==0.9.0
gunicorn==19.9.0